requirements.txt
django==4.1.13
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient